better storybook configuration
