+++ .storybook/preview-head.html
@@ -1,0 +1,4 @@
+<style id="app-styles"></style>
+<style id="app-fonts"></style>
+<link href="https://fonts.googleapis.com/css?family=Roboto:300,400,500" rel="stylesheet">
+<link href="https://maxcdn.bootstrapcdn.com/font-awesome/4.7.0/css/font-awesome.min.css" rel="stylesheet">
