+++ config/default.js
@@ -6,6 +6,8 @@ module.exports = {
     head: `
       <style id="app-styles"></style>
       <style id="app-fonts"></style>
+      <link href="https://cdnjs.cloudflare.com/ajax/libs/normalize/7.0.0/normalize.min.css" rel="stylesheet" />
+      <link href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/4.7.0/css/font-awesome.min.css" rel="stylesheet" />
     `,
     body: `<div id='app'></div>`,
   },
